+++ src/components/Login.js
@@ -1,5 +1,7 @@
+/* eslint-disable import/no-cycle */
 // eslint-disable-next-line import/no-cycle
-import { accesUser, accesGoogle/* accesFacebook */ } from '../lib/auth.js';
+import { accesUser, accesGoogle } from '../lib/auth.js';
+import { accesFacebook } from '../lib/controller-firebase/auth-functions.js';
 // eslint-disable-next-line import/no-cycle
 import { onNavigate } from '../Router.js';
 
@@ -74,7 +76,7 @@ export const Login = () => {
   icoFb.className = 'icon-fb';
   divIcon.appendChild(icoGoogle);
   divIcon.appendChild(icoFb);
-  /* icoFb.addEventListener('click', () => (accesFacebook()));*/
+  icoFb.addEventListener('click', () => (accesFacebook()));
 
   const divCardFooter = document.createElement('div');
   divCardFooter.className = 'text-center';
