+++ src/lib/auth.js
@@ -97,27 +97,6 @@ export async function accesGoogle() {
   return result;
 }
 
-/* // AUTENTICACIÓN CON FB
-const provider2 = new FacebookAuthProvider();
-
-export function accesFacebook() {
-  const auth = getAuth();
-  signInWithPopup(auth, provider2)
-    .then((result) => {
-      // The signed-in user info.
-      // This gives you a Facebook Access Token. You can use it to access the Facebook API.
-      const credential = FacebookAuthProvider.credentialFromResult(result);
-      const accessToken = credential.accessToken;
-      const user = result.user;
-      console.log(accessToken, user);
-      onNavigate('/feed');
-    })
-    .catch((error) => {
-      const credential = FacebookAuthProvider.credentialFromError(error);
-      console.log(credential, error);
-    });
-} */
-
 // REESTABLECER CONTRASEÑA
 export function restorePassword(email, message) {
   createNewPassword(email)
