+++ src/lib/controller-firebase/auth-functions.js
@@ -1,11 +1,12 @@
 /* eslint-disable no-unused-expressions */
 import {
   getAuth, createUserWithEmailAndPassword, signInWithEmailAndPassword, signInWithPopup,
-  GoogleAuthProvider, /* FacebookAuthProvider, */ signOut,
+  GoogleAuthProvider, FacebookAuthProvider, signOut,
   updatePassword, onAuthStateChanged, sendPasswordResetEmail,
 } from '../imports/firebase-imports.js';
 // eslint-disable-next-line import/no-cycle
 import { returnLogin } from '../auth.js';
+import { onNavigate } from '../../Router.js';
 
 export const createUser = (email, password) => {
   const auth = getAuth();
@@ -73,3 +74,24 @@ export const verifyUserActive = () => {
     }
   });
 };
+
+// AUTENTICACIÓN CON FB
+const provider2 = new FacebookAuthProvider();
+
+export function accesFacebook() {
+  const auth = getAuth();
+  signInWithPopup(auth, provider2)
+    .then((result) => {
+      // The signed-in user info.
+      // This gives you a Facebook Access Token. You can use it to access the Facebook API.
+      const credential = FacebookAuthProvider.credentialFromResult(result);
+      const accessToken = credential.accessToken;
+      const user = result.user;
+      console.log(accessToken, user);
+      onNavigate('/feed');
+    })
+    .catch((error) => {
+      const credential = FacebookAuthProvider.credentialFromError(error);
+      console.log(credential, error);
+    });
+}
